Listo. eliminar usuarios y admins finalizado. Además del middleware para control de usuarios de firebase y superusuarios.

+++ routes/web.php
@@ -4,6 +4,7 @@ use App\Http\Controllers\AdminController;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\VerificationController;
 use App\Http\Middleware\EnsureIsSuperAdmin;
+use App\Http\Middleware\FirebaseUser;
 use Illuminate\Support\Facades\Route;
 
 //Pantalla de inicio
@@ -16,25 +17,31 @@ Route::get('/', function () {
 Route::get('/email/verify', [VerificationController::class, 'verify'])->name('verify');
 Route::post('/admin/register', [AdminController::class, 'store'])->name('admin.store');
 Route::post('/admin/login', [AdminController::class, 'login'])->name('admin.login');
+Route::any('/logout', [AdminController::class, 'logout'])->name('logout');
 
-//Aqui solo van las rutas que ejerce el gerente/manager/superadmin, lo que sea
-Route::middleware([EnsureIsSuperAdmin::class])->group(function () {
+
+// Aplicar FirebaseUser Middleware a todas las rutas relevantes
+
+// Aquí solo van las rutas que ejerce el gerente/manager/superadmin, lo que sea
+Route::middleware([EnsureIsSuperAdmin::class, FirebaseUser::class])->group(function () {
     Route::get('/admins', [AdminController::class, 'index'])->name('manager.admin-list');
     Route::get('/admins/{admin}/edit', [AdminController::class, 'edit'])->name('manager.admin-edit');
-    Route::put('/admins/{admin}', [AdminController::class, 'update'])->name('manager.admin-update');  
-    Route::delete('/admins/{admin}', [AdminController::class, 'destroy'])->name('manager.admin-destroy'); 
+    Route::put('/admins/{admin}', [AdminController::class, 'update'])->name('manager.admin-update');
+    Route::delete('/admins/{admin}', [AdminController::class, 'destroy'])->name('manager.admin-destroy');
     Route::get('/actions', [AdminController::class, 'showActions'])->name('admin.actions');
     Route::post('/manager/activations/send/{id}', [AdminController::class, 'sendActivation'])->name('manager.activations.send');
     Route::get('/manager/activations', [AdminController::class, 'pendingActivation'])->name('manager.activations');
+
+
+    //Lista de usuarios.
+    Route::get('users', [UserController::class, 'index'])->name('admins.user-list')->withoutMiddleware([EnsureIsSuperAdmin::class]);
+    //Edición de usuarios.
+    Route::get('users/{user}/edit', [UserController::class, 'edit'])->name('admins.user-edit')->withoutMiddleware([EnsureIsSuperAdmin::class]);
+    Route::put('users/{user}', [UserController::class, 'update'])->name('users.update')->withoutMiddleware([EnsureIsSuperAdmin::class]);
+    //Eliminar usuario.
+    Route::delete('users/{user}', [UserController::class, 'destroy'])->name('admins.user-destroy')->withoutMiddleware([EnsureIsSuperAdmin::class]);
 });
 
-//Lista de usuarios.
-Route::get('users', [UserController::class, 'index'])->name('admins.user-list');
-//Edición de usuarios.
-Route::get('users/{user}/edit', [UserController::class, 'edit'])->name('admins.user-edit');
-Route::put('users/{user}', [UserController::class, 'update'])->name('users.update');
-//Eliminar usuario.
-Route::delete('users/{user}', [UserController::class, 'destroy'])->name('admins.user-destroy');
 
 // Llamar a esta URL en caso de que se elimine el admin principal.
 // Route::get('createSuperAdmin', [AdminController::class, 'createSuperAdmin']);
